Add files via upload
"The Indian Premier League (IPL) T20 cricket tournament is one of the most popular cricket leagues in the world. This repository contains an extensive Exploratory Data Analysis (EDA) of ball-by-ball data from multiple seasons of the IPL. The EDA provides valuable insights into the performance of teams and players by analyzing various aspects of the tournament.
The analysis begins with identifying the top players for categories such as player of the match, top run scorers, and provides visual representations of the results through graphs and charts. The EDA also includes a detailed analysis of the toss winner, first batting and winning teams, distribution of runs, winning team percentages, and the number of matches played in each city. The results provide an in-depth understanding of the trends and patterns in the IPL T20 data.
The data analysis is performed using Python  libraries, and the results are presented in Jupyter notebook with code and visualizations. The notebooks provide an interactive and comprehensive view of the data and the results of the analysis. The visualizations are created using popular data visualization libraries such as Matplotlib and Seaborn and help in making the results easy to understand and interpret.
The goal of this project is to provide valuable insights into the IPL T20 tournament and to help cricket enthusiasts and analysts to better understand the performance of teams and players. This repository is open to contributions, so feel free to fork the repository and submit pull requests. Whether you are a cricket fan, a data analyst, or a student of data science, this repository is a great resource to learn more about the IPL T20 tournament and the power of data analysis."